refactor: context api 적용
