+++ src/components/CardCompanySection/CardCompanySection.tsx
@@ -1,20 +1,17 @@
 import { CARD_COMPANY } from '../../constants';
-import Dropdown2 from '../Dropdown/Dropdown2';
+import { useFormContext } from '../../contexts/useFormContext';
+import Dropdown from '../Dropdown/Dropdown';
 import { InputSection } from '../InputSection/InputSection';
 
-export type CardCompanySectionProps = {
-  value: string;
-  onSelect: (value: string) => void;
-};
-
-export default function CardCompanySection({ value, onSelect }: CardCompanySectionProps) {
+export default function CardCompanySection() {
+  const { company, handleCompanySelect } = useFormContext();
   return (
     <div>
       <InputSection.TitleWrapper>
         <InputSection.Title title="카드사를 선택해 주세요" />
         <InputSection.SubTitle title="현재 국내 카드사만 가능합니다." />
       </InputSection.TitleWrapper>
-      <Dropdown2 placeholder="카드사를 선택해주세요" list={CARD_COMPANY} value={value} onSelect={onSelect} />
+      <Dropdown placeholder="카드사를 선택해주세요" list={CARD_COMPANY} value={company} onSelect={handleCompanySelect} />
     </div>
   );
 }
