+++ src/components/PasswordSection/PasswordSection.tsx
@@ -1,13 +1,10 @@
+import { useFormContext } from '../../contexts/useFormContext';
 import Input from '../Input/Input';
 import { InputSection } from '../InputSection/InputSection';
 import styles from './PasswordSection.module.css';
 
-export type PasswordSectionProps = {
-  password: { value: string; errorMessage: string };
-  handlePasswordChange: (value: string) => void;
-};
-
-export default function PasswordSection({ password, handlePasswordChange }: PasswordSectionProps) {
+export default function PasswordSection() {
+  const { handlePasswordChange, password } = useFormContext();
   return (
     <div className={styles.sectionContainer}>
       <InputSection.TitleWrapper>
